refactor: Implement BaseHoverPopup for ContextWindowPopup
